See merge request trixnity/trixnity!482

support files larger than 2.1GB

Closes #265

fixes #265

+++ trixnity-client/trixnity-client-repository-realm/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/realm/RealmMediaCacheMappingRepository.kt
@@ -13,7 +13,7 @@ internal class RealmMediaCacheMapping : RealmObject {
     @PrimaryKey
     var cacheUri: String = ""
     var mxcUri: String? = null
-    var size: Int? = null
+    var size: Long? = null
     var contentType: String? = null
 }
