+++ trixnity-core/src/commonMain/kotlin/net/folivo/trixnity/core/model/events/m/room/VideoInfo.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class VideoInfo(
     @SerialName("duration")
-    val duration: Int? = null,
+    val duration: Long? = null,
     @SerialName("h")
     val height: Int? = null,
     @SerialName("w")
@@ -14,7 +14,7 @@ data class VideoInfo(
     @SerialName("mimetype")
     override val mimeType: String? = null,
     @SerialName("size")
-    override val size: Int? = null,
+    override val size: Long? = null,
     @SerialName("thumbnail_url")
     val thumbnailUrl: String? = null,
     @SerialName("thumbnail_file")
